+++ src/RerollKnight/Assets/Code/Field/PickCellAsTargetSystem.cs
@@ -24,7 +24,6 @@ namespace Code
 			foreach (var e in entites)
 			{
 				e.Pick();
-				_contexts.ToGameState(WaitingForAbilityUsage);
 			}
 		}
 	}
